+++ src/param.rs
@@ -2,6 +2,7 @@ use std::fs::File;
 use std::io::{BufRead, BufReader};
 use serde::{Deserialize, Serialize};
 use std::error::Error;
+use crate::FloatType;
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Param {
@@ -30,11 +31,11 @@ pub struct Data {
     #[serde(default = "pvalue_method_default")]                      // Path to y data
     pub pvalue_method: String,
     #[serde(default = "feature_minimal_prevalence_pct_default")]                      // Path to y data
-    pub feature_minimal_prevalence_pct: f64, // Minimum prevalence
+    pub feature_minimal_prevalence_pct: FloatType, // Minimum prevalence
     #[serde(default = "feature_maximal_pvalue_default")]                      // Path to y data
-    pub feature_maximal_pvalue: f64, // Minimum prevalence
+    pub feature_maximal_pvalue: FloatType, // Minimum prevalence
     #[serde(default = "feature_minimal_feature_value_default")]                      // Path to y data
-    pub feature_minimal_feature_value: f64, // Minimum prevalence
+    pub feature_minimal_feature_value: FloatType, // Minimum prevalence
 }
 
 #[derive(Debug, Serialize, Deserialize)]
@@ -47,12 +48,12 @@ pub struct GA {
     pub max_age_best_model: usize,                 // Stop if the best model has not change for this long
     //pub kmin: u32,                           // Minimum value of k
     //pub kmax: u32,                           // Maximum value of k
-    pub kpenalty: f64,                       // A penalty of this per value of k is deleted from AUC in the fit function
-    pub select_elite_pct: f64,               // Elite selection percentage
-    pub select_random_pct: f64,              // Random selection percentage
-    pub mutated_children_pct: f64,        // Mutated individuals percentage
-    pub mutated_features_pct: f64,           // Mutated features percentage
-    pub mutation_non_null_chance_pct: f64,    // Chance pct that a mutation gives an non null value
+    pub kpenalty: FloatType,                       // A penalty of this per value of k is deleted from AUC in the fit function
+    pub select_elite_pct: FloatType,               // Elite selection percentage
+    pub select_random_pct: FloatType,              // Random selection percentage
+    pub mutated_children_pct: FloatType,        // Mutated individuals percentage
+    pub mutated_features_pct: FloatType,           // Mutated features percentage
+    pub mutation_non_null_chance_pct: FloatType,    // Chance pct that a mutation gives an non null value
     #[serde(default = "feature_importance_permutations_default")]    
     pub feature_importance_permutations: usize,
 }
@@ -73,8 +74,8 @@ fn min_epochs_default() -> usize { 10 }
 fn max_age_best_model_default() -> usize { 10 }
 fn algorithm_default() -> String { "ga".to_string() }
 fn pvalue_method_default() -> String { "studentt".to_string() }
-fn feature_minimal_prevalence_pct_default() -> f64 { 10.0 }
-fn feature_maximal_pvalue_default() -> f64 { 0.5 }
+fn feature_minimal_prevalence_pct_default() -> FloatType { 10.0 }
+fn feature_maximal_pvalue_default() -> FloatType { 0.5 }
 fn feature_importance_permutations_default() -> usize { 10 }
-fn feature_minimal_feature_value_default() -> f64 { 0.0 }
+fn feature_minimal_feature_value_default() -> FloatType { 0.0 }
 fn thread_number_default() -> usize { 1 }
